added and modified some GroupRecipeBook classes
