+++ src/main/java/UserInterface/CLI/PageManager.java
@@ -4,17 +4,15 @@ import main.java.UserInterface.Commands.AdminCommands.CreateAccountCommand;
 import main.java.UserInterface.Commands.AdminCommands.SignInCommand;
 import main.java.UserInterface.Commands.AdminCommands.SignOutCommand;
 import main.java.UserInterface.Commands.Command;
+import main.java.UserInterface.Commands.GeneralRecipeBookCommands.GroupRecipeBookCommands.*;
+import main.java.UserInterface.Commands.GeneralRecipeBookCommands.RecipeBookCommands.*;
 import main.java.UserInterface.Commands.GoBackCommand;
 import main.java.UserInterface.Commands.GroupCommands.AddGroupMemberCommand;
 import main.java.UserInterface.Commands.GroupCommands.CreateGroupCommand;
 import main.java.UserInterface.Commands.GroupCommands.EnterManageGroupCommand;
 import main.java.UserInterface.Commands.GroupCommands.RemoveGroupMemberCommand;
 import main.java.UserInterface.Commands.GroupCommands.ViewGroupsCommand;
-import main.java.UserInterface.Commands.GroupRecipeBookCommands.*;
-import main.java.UserInterface.Commands.GroupRecipeBookCommands.EnterGroupRecipeBookCommand;
-import main.java.UserInterface.Commands.GroupRecipeBookCommands.AddGroupSubRecipeBookCommand;
 //import main.java.UserInterface.Commands.GroupRecipeViewerCommands.GetNewGroupRecipeCommand; //TODO: check
-import main.java.UserInterface.Commands.RecipeBookCommands.*;
 import main.java.UserInterface.Commands.RecipeViewerCommands.EnterRecipeViewerCommand;
 import main.java.UserInterface.Commands.RecipeViewerCommands.FilterAndSortCommands.FilterRecipeBookCommand;
 import main.java.UserInterface.Commands.RecipeViewerCommands.FilterAndSortCommands.RemoveFilterCommand;
