+++ src/main/java/UserInterface/Commands/GeneralRecipeBookCommands/GroupRecipeBookCommands/AddGroupSubRecipeBookCommand.java
@@ -1,4 +1,4 @@
-package main.java.UserInterface.Commands.GroupRecipeBookCommands;
+package main.java.UserInterface.Commands.GeneralRecipeBookCommands.GroupRecipeBookCommands;
 
 import main.java.Entities.User;
 import main.java.Entities.Group;
