+++ src/main/java/UserInterface/Commands/GeneralRecipeBookCommands/GroupRecipeBookCommands/EnterGroupSubRecipeBookCommand.java
@@ -1,4 +1,4 @@
-package main.java.UserInterface.Commands.GroupRecipeBookCommands;
+package main.java.UserInterface.Commands.GeneralRecipeBookCommands.GroupRecipeBookCommands;
 
 import main.java.UseCases.GroupRecipeBookManager;
 import main.java.UserInterface.Commands.Command;
