+++ src/main/java/UserInterface/Commands/GeneralRecipeBookCommands/GroupRecipeBookCommands/ListGroupRecipeBooksCommand.java
@@ -1,4 +1,4 @@
-package main.java.UserInterface.Commands.GroupRecipeBookCommands;
+package main.java.UserInterface.Commands.GeneralRecipeBookCommands.GroupRecipeBookCommands;
 import main.java.Entities.GroupRecipeBook;
 import main.java.Gateways.GroupRecipeBookCSVReader;
 import main.java.UserInterface.Commands.Command;
