+++ src/main/java/UserInterface/Commands/GeneralRecipeBookCommands/GroupRecipeBookCommands/RemoveGroupRecipeCommand.java
@@ -1,4 +1,4 @@
-package main.java.UserInterface.Commands.GroupRecipeBookCommands;
+package main.java.UserInterface.Commands.GeneralRecipeBookCommands.GroupRecipeBookCommands;
 
 import main.java.UseCases.GroupSubRecipeBookManager;
 import main.java.Gateways.GroupRecipeBookCSVReader;
