+++ src/main/java/UserInterface/Commands/GeneralRecipeBookCommands/RecipeBookCommands/EnterRecipeBookCommand.java
@@ -1,0 +1,18 @@
+package main.java.UserInterface.Commands.GeneralRecipeBookCommands.RecipeBookCommands;
+
+import main.java.UserInterface.Commands.Command;
+import main.java.UserInterface.Commands.HelpCommand;
+import main.java.UserInterface.UserInterface;
+
+public class EnterRecipeBookCommand extends Command {
+    public EnterRecipeBookCommand() {
+        super("enter recipe book", "Enters the recipe book");
+    }
+
+    @Override
+    public void runAction(UserInterface UI) {
+        UI.getPageManager().enterRecipeBook();
+        Command help = new HelpCommand();
+        help.runAction(UI);
+    }
+}
