--- src/main/java/UserInterface/Commands/GroupRecipeBookCommands/ListGroupSubRecipeBookRecipesCommand.java
@@ -1,4 +1,0 @@
-package main.java.UserInterface.Commands.GroupRecipeBookCommands;
-import main.java.UserInterface.Commands.Command;
-
-//TODO: public class ListGroupSubRecipeBookRecipesCommand extends Command{}
