+++ src/main/java/UserInterface/GUI/GUIPages/RecipeBook.java
@@ -1,7 +1,7 @@
 package main.java.UserInterface.GUI.GUIPages;
 
-import main.java.UserInterface.Commands.RecipeBookCommands.ListRecipeBookCommand;
-import main.java.UserInterface.Commands.RecipeBookCommands.RemoveRecipeCommand;
+import main.java.UserInterface.Commands.GeneralRecipeBookCommands.RecipeBookCommands.ListRecipeBookCommand;
+import main.java.UserInterface.Commands.GeneralRecipeBookCommands.RecipeBookCommands.RemoveRecipeCommand;
 import main.java.UserInterface.GUI.GUIForm;
 
 import javax.swing.JFrame;
